build: error in the provider viewContext refactor
