+++ src/components/admin/Skill/TableSkills/index.tsx
@@ -1,19 +1,18 @@
-import React, { useContext } from "react";
+import React from "react";
 import usePagination from "../../../../hooks/usePagination";
 import useSkills from "../../../../hooks/useSkills";
 import FormSkill from "../FormSkill";
-import MessageForm from "../../../ui/MessageForm";
 import BaseTableForm from "../../../ui/BaseTableForm";
 import ContentTableForm from "../../../ui/BaseTableForm/ContentTableForm";
 import ElementOrForm from "../../../ui/ElementOrForm";
-import ViewContext from "../../../../providers/viewContext";
+import useView from "../../../../providers/view/viewContext";
 
 interface TableSkillsProps {
   data: any;
   title?: string;
 }
 export default function TableSkills(props: TableSkillsProps) {
-  const { view, setView } = useContext(ViewContext);
+  const { view, setView } = useView();
   const { pagination, botaoMostrarMais } = usePagination(3, 3);
   const {
     setType,
@@ -40,7 +39,7 @@ export default function TableSkills(props: TableSkillsProps) {
                   setType(element.type);
                   setSkill(element.skill);
                   setIdSkill(element._id);
-                  setView(true)
+                  setView(true);
                 }}
                 onClickTrash={() => deleteSkill(element._id)}
               />
