+++ src/pages/api/academy/update/index.ts
@@ -1,5 +1,4 @@
 import AcademyController from "../../../../controllers/AcademyController";
-import { adminAccessApi } from "../../../../back-end/security/admin.acess";
 import { apiCallPatternFormat } from "../../../../patterns/api/apiCall";
 import { academyUpdateValidate } from "../../../../middlewares/academy/updateValidate";
