+++ src/patterns/api/apiCall.ts
@@ -1,26 +1,26 @@
 import Database from "../../infra/db";
 import { requestFormater } from "./requestFormater";
 
+interface entitiePatternApi {
+  validate: (validate, validationExpress) => void;
+  controller: (req, res) => void;
+  method: string;
+}
+
 export async function apiCallPatternFormat(
   req,
   res,
-  entitie: {
-    validate: (validate, validationExpress) => void;
-    controller: (req, res) => void;
-    method: string;
-  }
+  entitie: entitiePatternApi
 ) {
   if (!req.body && entitie.method != "GET") {
-    res
-      .status(404)
-      .json({
-        errors: ["body é requisitado e não foi encontrado na requisição."],
-      });
+    res.status(404).json({
+      errors: ["body é requisitado e não foi encontrado na requisição."],
+    });
     return;
   }
 
-  const formatedRequest = await requestFormater(req);
-  await entitie.validate(formatedRequest, res);
+  const formatRequestStringToJson = await requestFormater(req);
+  await entitie.validate(formatRequestStringToJson, res);
 
   await responseAPI(
     {
@@ -28,7 +28,7 @@ export async function apiCallPatternFormat(
       requestedMethod: req.method,
       response: res,
     },
-    entitie.controller(formatedRequest, res)
+    entitie.controller(formatRequestStringToJson, res)
   );
 }
