--- src/providers/view/view.tsx
@@ -1,6 +1,0 @@
-import { useContext } from "react";
-import { ViewContext } from "./viewContext";
-
-export default function useView() {
-  return useContext(ViewContext);
-}
